Time: 7 ms (83.56%), Space: 47.6 MB (61.57%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.java
@@ -8,7 +8,7 @@ class Solution {
 
         while(left < right){
             int mid = left + (right - left) / 2;
-            long hours = 0;
+            int hours = 0;
             for(int pile : piles){
                 hours += (pile + mid - 1) / mid;
             }
